Switched to a template for the stringToAsset function
also renamed files but have to do it in 2 commits because git ignores file case
