+++ NorthstarDedicatedTest/NorthstarDedicatedTest.vcxproj
@@ -121,7 +121,7 @@
     <ClInclude Include="ns_version.h" />
     <ClInclude Include="plugins.h" />
     <ClInclude Include="plugin_abi.h" />
-    <ClInclude Include="scriptUtility.h" />
+    <ClInclude Include="scriptutility.h" />
     <ClInclude Include="serverchathooks.h" />
     <ClInclude Include="clientauthhooks.h" />
     <ClInclude Include="color.h" />
@@ -617,7 +617,7 @@
     <ClCompile Include="scriptmodmenu.cpp" />
     <ClCompile Include="scriptserverbrowser.cpp" />
     <ClCompile Include="scriptsrson.cpp" />
-    <ClCompile Include="scriptUtility.cpp" />
+    <ClCompile Include="scriptutility.cpp" />
     <ClCompile Include="serverauthentication.cpp" />
     <ClCompile Include="miscserverscript.cpp" />
     <ClCompile Include="serverchathooks.cpp" />
