+++ NorthstarDedicatedTest/NorthstarDedicatedTest.vcxproj.filters
@@ -1515,7 +1515,7 @@
     <ClInclude Include="NSMem.h">
       <Filter>Header Files\Shared\ExploitFixes</Filter>
     </ClInclude>
-    <ClInclude Include="scriptUtility.h">
+    <ClInclude Include="scriptutility.h">
       <Filter>Header Files\Shared</Filter>
     </ClInclude>
   </ItemGroup>
@@ -1685,7 +1685,7 @@
     <ClCompile Include="clientruihooks.cpp">
       <Filter>Source Files\Client</Filter>
     </ClCompile>
-    <ClCompile Include="scriptUtility.cpp">
+    <ClCompile Include="scriptutility.cpp">
       <Filter>Source Files\Shared</Filter>
     </ClCompile>
   </ItemGroup>
